extract base tests
